Added PersonDaoImpl and PersonDaoImplTests, modified Person to make it compatible with JodaTime serialisation, minor bug fixes.

+++ src/main/java/com/rms/pilotapi/dao/PersonDao.java
@@ -3,11 +3,11 @@ package com.rms.pilotapi.dao;
 import com.rms.pilotapi.core.Person;
 
 public interface PersonDao {
-    Person getPerson(Integer id);
+    Person getPerson(long id);
 
     Person createPerson(Person person);
 
-    Person updatePerson(Integer id, Person person);
+    Person updatePerson(long id, Person person);
 
-    boolean deletePerson(Integer id);
+    boolean deletePerson(long id);
 }
